+++ app/src/main/res/drawable/button_shape.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <shape xmlns:android="http://schemas.android.com/apk/res/android">
     <solid
-        android:color="@color/gray"
+        android:color="@color/gray1"
         />
     <corners
         android:radius="10dp"/>
