+++ app/src/main/res/values/colors.xml
@@ -1,10 +1,12 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="black">#FF000000</color>
+    <color name="black">#1E1E1E</color>
     <color name="white">#FFFFFFFF</color>
     <color name="lightPurple">#6071CC</color>
-    <color name="gray">#2E2E2E</color>
+    <color name="gray1">#2E2E2E</color>
+    <color name="gray2">#5f5f5f</color>
     <color name="transparent">#00000000</color>
     <color name="YellowLight">#FCF8CE</color>
+    <color name="red">#FF2828</color>
 
 </resources>
